Remove ComponentId from ScryptoActor
