+++ radix-engine/tests/blueprints/Cargo.toml
@@ -1,6 +1,7 @@
 [workspace]
 members = [
     "abi",
+    "access_rules",
     "arguments",
     "bucket",
     "component",
