+++ radix-engine/tests/blueprints/access_rules/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "access_rules"
+version = "0.7.0"
+edition = "2021"
+
+[dependencies]
+sbor = { path = "../../../../sbor" }
+scrypto = { path = "../../../../scrypto" }
+
+[dev-dependencies]
+radix-engine = { path = "../../../../radix-engine" }
+
+[lib]
+crate-type = ["cdylib", "lib"]
