+++ radix-engine/tests/blueprints/component/src/lib.rs
@@ -4,4 +4,3 @@ pub mod chess;
 pub mod component;
 pub mod cross_component;
 pub mod external_blueprint_target;
-pub mod mutable_access_rules;
