+++ radix-engine/tests/reentrancy.rs
@@ -1,9 +1,9 @@
-use radix_engine::engine::{ KernelError, LockState, RuntimeError, TrackError };
+use radix_engine::engine::{KernelError, LockState, RuntimeError, TrackError};
 use radix_engine::ledger::TypedInMemorySubstateStore;
 use radix_engine::types::*;
 use radix_engine_interface::api::types::RENodeId;
 use radix_engine_interface::core::NetworkDefinition;
-use radix_engine_interface::{data::*};
+use radix_engine_interface::data::*;
 use scrypto_unit::*;
 use transaction::builder::ManifestBuilder;
 
@@ -94,7 +94,11 @@ fn read_then_mut_reentrancy_should_not_be_possible() {
     // Act
     let manifest = ManifestBuilder::new(&NetworkDefinition::simulator())
         .lock_fee(FAUCET_COMPONENT, 10u32.into())
-        .call_method(component_address, "call_mut_self_2", args!(component_address))
+        .call_method(
+            component_address,
+            "call_mut_self_2",
+            args!(component_address),
+        )
         .build();
     let receipt = test_runner.execute_manifest(manifest, vec![]);
